+++ NEWS.md
@@ -12,6 +12,7 @@ editor_options:
 -   Support `robustbase::lmrob()` in `stat_poly_line()` and `stat_poly_eq()`, polynomials fitted by robust regression methods. 
 -   Support `nlme::gls()`, `MASS::lqs()`, `robustbase::lmrob()` and model fit objects of classes for which method `weights()` is not available in `stat_fit_deviations()`, `stat_fit_fitted()`, `stat_fit_residuals()`.
 -   Update `scale_x_logFC()`and `scale_y_logFC()` for improved compatibility with 'ggplot2' (== 3.5.1.9000).
+-   Code breaking change in `stat_fit_deviations()`: consistently return prior weights in variable `weights` and new variable `robustness.weights` for the implicit weights used by robust fit methods.
 -   Bug fix: the value of 'method.label' not always set.
 -   Add HTML Cheat Sheet.
 -   Add article 'Custom polynomial models' as on-line only documentation.